[Chore] loader, set, model, trainer에 대한 base 생성
- 각각의 경우에 대한 base class 생성
- 해당 class를 상속하여 추후 여러가지 구현
- 현재는 빈 깡통으로 만들어놓은 상태
